requirements.txt
streamlit==1.36.0
pydantic==2.9.2
pydantic_core==2.23.4

langchain==0.3.1
langchain-core==0.3.6
langchain-community==0.3.1
langchain-chroma == 0.1.4

sentence_transformers==2.2.2
instructorembedding==1.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
streamlit-extras==0.4.3
pillow==10.4.0
sseclient-py==1.8.0
# unstructured==0.14.9
# unstructured_inference==0.7.36
# unstructured_pytesseract==0.3.12
# pytesseract==0.3.10
chromadb==0.5.3
langgraph==0.2.28
openpyxl==3.1.4
psutil==6.0.0
pillow_heif==0.16.0
ipython==8.26.0
# PyMuPDF==1.23.4
# PyMuPDFb==1.23.3

#LLM Eval
weave==0.51.1